\name{csDEX-package}
\alias{csDEX-package}
\alias{csDEX}
\docType{package}
\title{
\packageTitle{csDEX}
}
\description{
\packageDescription{csDEX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{csDEX}
\packageIndices{csDEX}
csDEX (Condition-specific Differential Exon Expression) is a family of general
linear models (GLMs), used to model differential splicing given a set of
sequencing experiments based on RNA-seq. The methods assume tens to hundreds
experimental conditions to be compared, and the result are
changes in usage of splice sites that are specific to an
experimental condition. 

Capabilities:
  \item{handling of sequence data in form of read counts or percentages (e.g. percent-spliced in, PSI),}
  \item{modelling all types of splicing events,}
  \item{discovery and ranking of condition-specific changes,}
  \item{parallel processing of multiple genes (groups),}
  \item{time-efficient model approximation with no change in false positive rate.}
}
\author{
\packageAuthor{csDEX}

Maintainer: \packageMaintainer{csDEX}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
}
