\name{csDEX-package}
\alias{csDEX-package}
\alias{csDEX}
\docType{package}
\title{
\packageTitle{csDEX}
}
\description{
\packageDescription{csDEX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{csDEX}
\packageIndices{csDEX}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{csDEX}

Maintainer: \packageMaintainer{csDEX}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
}
